moved signal handling tests to AbstractLoopTest
do not merge yet, this currently fails on various combinations of php versions,
I'd like to fix it after #45 is merged to have tests
run against all systems.
